Restrict allowed patterns for character
Right now %|TEST| would be allowed.
This PR does not fix `%TEST` or %test which is still allowed (false-positive).
It should only allow single-characters or specific characters like
`%space`, `%tab`, `%newline` etc. but that would create a lot of complexity.
`%test` wouldn't compile in Magik so IMO it is okay to keep that false-positive.